Change Zendesk case and URL
The official case for this company name is Zendesk, and they are an
American company. Their main site is zendesk.com, although there is a
geo-based redirection to zendesk.co.uk if the user is from the U.K.

This corrects the case from ZenDesk to Zendesk, and changes the URL to
point to their main site .com, instead of the U.K. version.